+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Modal
   * Changes will also effect Dialog, Sidebar etc...
 
 Dialog
+
   * New props `showCloseIcon` (defaulted to true) which show and hides the close icon
 
 ## Promises
@@ -47,6 +48,12 @@ Promises Polyfill. Carbon now contains a ES6 Promises helper which can be import
 ```javascript
   import from 'carbon/lib/utils/promises';
 ```
+
+## Notifications Updates
+
+Message
+
+  * New props `transparent` (defaulted to false) which if set to true sets the background to transparent
 
 # 0.17.1